clion refactoring

+++ string_buffer.c
@@ -1,22 +1,20 @@
 #include <stdlib.h>
-#include <string.h>
 #include <stdbool.h>
 #include <stdio.h>
-#include <ctype.h>
 #include "string_buffer.h"
 
-void buffer_init(sbuffer * b, int maxlen) {
+void buffer_init(sbuffer *b, size_t maxlen) {
     b->content = malloc(sizeof(char) * (maxlen + 1));
     b->content[0] = '\0';
     b->maxlen = maxlen;
     b->len = 0;
 }
 
-void buffer_teardown(sbuffer * b) {
+void buffer_teardown(sbuffer *b) {
     free(b->content);
 }
 
-bool buffer_putc(sbuffer * b, char c){
+bool buffer_putc(sbuffer *b, char c) {
     if (b->len >= b->maxlen)
         return false;
     b->content[b->len] = c;
@@ -25,14 +23,14 @@ bool buffer_putc(sbuffer * b, char c){
     return true;
 }
 
-void buffer_write(sbuffer * b, FILE * f){  
+void buffer_write(sbuffer *b, FILE *f) {
     fwrite(b->content,
             sizeof(char),
             b->len,
             f);
 }
 
-void buffer_clear(sbuffer * b){
+void buffer_clear(sbuffer *b) {
     b->content[0] = '\0';
     b->len = 0;
 }
